Update Makevars.win
By changing hard-coded ar to variable, stringi can inherit any prefixes which may be added to R when compiled from source. This will be useful for the experimental 4.9.2 toolchain, which when link-time optimization (LTO) is being used, needs to call `ar`, `nm`, and `ranlib` by `gcc-ar`, `gcc-nm`, and `gcc-ranlib`. Changes to Mkrules and Makeconf in R are being investigated. Under the current toolset, or without LTO, $(AR) will just be `ar`.

Former-commit-id: 80488f8e1c73fd077e224b298c59aa81a137ebbb